rename resettable to counter.rs
