+++ src/common/metrics/src/counter.rs
